requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
PyJWT==2.8.0
cryptography==41.0.7
ldap3==2.9.1
python-dotenv==1.0.0
python-multipart==0.0.6 